pyproject.toml
[project]
name = "ai-mcp-news-search"
version = "0.6.0"
description = "Enterprise AI Platform — News Search MCP (Tavily-backed)"
requires-python = ">=3.11"
dependencies = [
    "enterprise-ai-platform-sdk @ git+ssh://git@github.com/narisun/ai-platform-sdk.git@v0.6.0",
    "tavily-python>=0.3.0,<1.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "unit: fast unit tests (no I/O, no Docker)",
    "component: multi-class tests with faked external systems",
    "integration: docker-compose stack required",
]
